+++ src/SWP/Bundle/SettingsBundle/Manager/SettingsManagerInterface.php
@@ -71,9 +71,17 @@ interface SettingsManagerInterface
     public function clearAllByScope(string $scope = ScopeContextInterface::SCOPE_GLOBAL): void;
 
     /**
-     * @param string $scope
+     * @param string                 $scope
+     * @param SettingsOwnerInterface $settingsOwner
      *
      * @return array
      */
-    public function getAllByScope(string $scope): array;
+    public function getByScopeAndOwner(string $scope, SettingsOwnerInterface $settingsOwner): array;
+
+    /**
+     * @param string $name
+     *
+     * @return array|null
+     */
+    public function getOneSettingByName(string $name): ?array;
 }